delete test

+++ app/Services/TaskService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Models\Task;
 use App\Repositories\Contracts\TaskRepositoryContract;
+use Illuminate\Support\Facades\Gate;
 
 class TaskService
 {
@@ -23,11 +24,20 @@ class TaskService
 
     public function updateTask(array $data, string $id): Task
     {
+
+        $task = $this->getTask($id);
+
+        Gate::authorize('update', $task);
+
         return $this->taskRepository->update($data, $id);
     }
 
     public function deleteTask(string $id)
     {
+        $task = $this->getTask($id);
+
+        Gate::authorize('delete', $task);
+
         return $this->taskRepository->delete($id);
     }
 }
